server/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pillow==10.2.0
python-dotenv==1.0.1
sqlalchemy==2.0.27
aiosqlite==0.19.0 